move logout to separate component
